Conflicts:
	crawl-ref/source/command.cc
	crawl-ref/source/dat/database/FAQ.txt
	crawl-ref/source/dat/database/monspell.txt
	crawl-ref/source/dat/descript/monsters.txt
	crawl-ref/source/describe.cc
	crawl-ref/source/evoke.cc
	crawl-ref/source/food.cc
	crawl-ref/source/invent.cc
	crawl-ref/source/item_use.cc
	crawl-ref/source/itemname.cc
	crawl-ref/source/itemprop.cc
	crawl-ref/source/items.cc
	crawl-ref/source/main.cc
	crawl-ref/source/melee_attack.cc
	crawl-ref/source/mon-act.cc
	crawl-ref/source/player.cc
	crawl-ref/source/quiver.cc
	crawl-ref/source/religion.cc
	crawl-ref/source/shopping.cc
	crawl-ref/source/spl-data.h
	crawl-ref/source/spl-other.cc
	crawl-ref/source/spl-transloc.cc
	crawl-ref/source/status.cc
	crawl-ref/source/teleport.cc
	crawl-ref/source/transform.cc

+++ crawl-ref/source/mon-place.cc
@@ -2475,7 +2475,7 @@ static band_type _choose_band(monster_type mon_type, int &band_size,
 
     case MONS_IRONHEART_PRESERVER:
         natural_leader = true;
-        switch(random2(3))
+        switch (random2(3))
         {
             case 0:
                 band = BAND_DEEP_ELF_HIGH_PRIEST;
@@ -3606,7 +3606,7 @@ static void _get_vault_mon_list(vector<mons_spec> &list)
         mons_spec spec;
 
 #if TAG_MAJOR_VERSION == 34
-        if (type == -1)
+        if ((int)type == -1)
             type = MONS_NO_MONSTER;
 #endif
         if (type == MONS_NO_MONSTER && base != MONS_NO_MONSTER)
